+++ core-ng/src/main/java/core/framework/impl/http/RetryHandler.java
@@ -3,6 +3,7 @@ package core.framework.impl.http;
 import core.framework.http.HTTPMethod;
 import core.framework.log.Markers;
 import core.framework.util.Threads;
+import org.apache.http.NoHttpResponseException;
 import org.apache.http.client.HttpRequestRetryHandler;
 import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.client.protocol.HttpClientContext;
@@ -34,7 +35,7 @@ public class RetryHandler implements HttpRequestRetryHandler {
         HttpClientContext clientContext = HttpClientContext.adapt(context);
         HttpUriRequest request = (HttpUriRequest) clientContext.getRequest();
 
-        if (retry(request, clientContext)) {
+        if (retry(request, clientContext, exception)) {
             logger.warn(Markers.errorCode("HTTP_COMMUNICATION_FAILED"), "http connection failed, retry soon", exception);
             Duration waitTime = waitTime(executionCount);
             waitBeforeRetry(waitTime);
@@ -44,10 +45,15 @@ public class RetryHandler implements HttpRequestRetryHandler {
         return false;
     }
 
-    boolean retry(HttpUriRequest request, HttpClientContext clientContext) {
+    boolean retry(HttpUriRequest request, HttpClientContext clientContext, IOException exception) {
         if (request.isAborted()) {
             return false;
         }
+
+        // with keep-alive + graceful shutdown, it's probably ok to retry on this exception even with non-idempotent methods
+        // this exception means server side drops the connection, the real case is during deployment (kube), persistent connection established via kube-proxy, and wouldn't know if old pod is deleted
+        if (exception instanceof NoHttpResponseException) return true;
+
         return !clientContext.isRequestSent() || idempotentMethods.contains(request.getMethod());
     }